+++ merqube_client_lib/api_client/merqube_client.py
@@ -1,15 +1,14 @@
 """
 Combined API Client for all merqube APIs
 """
-from typing import Any
+from typing import Any, cast
 
 import pandas as pd
 from cachetools import LRUCache, cached
 
-from merqube_client_lib.api_client.base import MerqubeApiClientBase
 from merqube_client_lib.api_client.indexapi import IndexAPIClient
+from merqube_client_lib.api_client.secapi import SecAPIClient
 from merqube_client_lib.pydantic_types import IndexDefinitionPatchPutGet as Index
-from merqube_client_lib.secapi.client import SecAPIClient
 from merqube_client_lib.session import MerqubeAPISession
 from merqube_client_lib.types import Manifest
 
@@ -22,16 +21,33 @@ class MerqubeAPIClient(IndexAPIClient, SecAPIClient):
     """
 
 
-class MerqubeAPIClientSingleIndex(MerqubeApiClientBase):
+class MerqubeAPIClientSingleIndex(MerqubeAPIClient):
     """
     Class that contains methods that deal with a single index
+    Instantiate with an index name ("name" field in the index manifest) and indicate if it is a realtime index
     """
 
-    def __init__(self, index_name: str, *args: Any, **kwargs: Any) -> None:
+    def __init__(self, index_name: str, is_intraday: bool = False, *args: Any, **kwargs: Any) -> None:
         super().__init__(*args, **kwargs)
 
         res = self.session.get_collection_single(f"/index?names={index_name}")
         self.index_id = res["id"]
+        self.is_intraday = is_intraday
+
+        mod = self.get_index_model()
+
+        # intraday model has a nasty type of [None | Intraday | Bool ..]
+        self._has_intraday = mod.intraday is not None and (
+            (isinstance(mod.intraday, bool) and mod.intraday) or mod.intraday.enabled is True
+        )
+
+        # get the security id for this index
+        self._sec_id = self.session.get_collection_single(f"/security/index?names={index_name}")["id"]
+        self._intra_sec_id = (
+            self.session.get_collection_single(f"/security/intraday_index?names={index_name}")["id"]
+            if self._has_intraday
+            else None
+        )
 
     def get_index_manifest(self) -> Manifest:
         """
@@ -45,21 +61,45 @@ class MerqubeAPIClientSingleIndex(MerqubeApiClientBase):
         """
         return Index.parse_obj(self.get_index_manifest())
 
-    def get_returns(
-        self,
-        returns_metric: str = "price_return",
-        start_date: pd.Timestamp | None = None,
-        end_date: pd.Timestamp | None = None,
-    ) -> pd.DataFrame:
-        raise NotImplementedError((returns_metric, start_date, end_date))
-
     def get_metrics(
         self,
         metrics: list[str],
+        use_intraday_metrics: bool = False,
         start_date: pd.Timestamp | None = None,
         end_date: pd.Timestamp | None = None,
     ) -> pd.DataFrame:
-        raise NotImplementedError((metrics, start_date, end_date))
+        """
+        Get a list of metrics for this index
+        Some indices are both end of day and intraday - use the flag to query for intraday metrics
+        """
+        return self.get_security_metrics(
+            sec_type="intraday_index" if use_intraday_metrics else "index",
+            sec_ids=[cast(str, self._intra_sec_id) if use_intraday_metrics else self._sec_id],
+            metrics=metrics,
+            start_date=start_date,
+            end_date=end_date,
+        )
+
+    def get_returns(
+        self,
+        returns_metric: str = "price_return",
+        use_intraday_metrics: bool = False,
+        start_date: pd.Timestamp | None = None,
+        end_date: pd.Timestamp | None = None,
+    ) -> pd.DataFrame:
+        """
+        Get returns for this index
+        Some indices are both end of day and intraday - use the flag to query for intraday metrics
+        """
+        if use_intraday_metrics and not self._has_intraday:
+            raise ValueError("This index is not an intraday index")
+
+        return self.get_metrics(
+            metrics=[returns_metric],
+            use_intraday_metrics=use_intraday_metrics,
+            start_date=start_date,
+            end_date=end_date,
+        )
 
     def get_portfolio(self) -> list[dict[str, Any]]:
         """
@@ -111,9 +151,19 @@ client_cache: LRUCache = LRUCache(maxsize=256)  # type: ignore
 
 @cached(cache=client_cache)
 def get_client(
-    user_session: MerqubeAPISession | None = None, token: str | None = None, **session_kwargs: Any
-) -> MerqubeAPIClient:
+    index_name: str | None = None,
+    is_intraday: bool = False,
+    user_session: MerqubeAPISession | None = None,
+    token: str | None = None,
+    **session_kwargs: Any,
+) -> MerqubeAPIClient | MerqubeAPIClientSingleIndex:
     """
-    Cached; returns a Merqube client for token
+    Cached; returns a Merqube client (or SingleClient if provided an index name) for token
     """
-    return MerqubeAPIClient(user_session=user_session, token=token, **session_kwargs)
+    return (
+        MerqubeAPIClientSingleIndex(
+            index_name=index_name, is_intraday=is_intraday, user_session=user_session, token=token, **session_kwargs
+        )
+        if index_name
+        else MerqubeAPIClient(user_session=user_session, token=token, **session_kwargs)
+    )